SolarScoreModel/requirements.txt
pandas>=2.0
numpy>=1.24
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
scipy>=1.11
scikit-learn>=1.4
xgboost>=2.0
matplotlib>=3.8
requests>=2.31
pyarrow>=14.0
psycopg2-binary>=2.9
joblib>=1.3
